Move network modules into crmprtd.networks
